src/worker_1/requirements.txt
python-telegram-bot[rate-limiter]==21.0.1
PyYAML==6.0
pymongo==4.6.2
python-dotenv==0.21.0
requests==2.31.0
motor==3.3.2